+++ contrib/platform/src/com/sun/jna/platform/win32/WinNT.java
@@ -2016,6 +2016,23 @@ public interface WinNT extends WinError, WinDef, WinBase, BaseTSD {
     int VER_PLATFORM_WIN32_WINDOWS = 1;
     int VER_PLATFORM_WIN32_NT = 2;
 
+    /*
+     * WIN32_WINNT version constants
+     */
+    short WIN32_WINNT_NT4 = 0x0400; // Windows NT 4.0
+    short WIN32_WINNT_WIN2K = 0x0500; // Windows 2000
+    short WIN32_WINNT_WINXP = 0x0501; // Windows XP
+    short WIN32_WINNT_WS03 = 0x0502; // Windows Server 2003
+    short WIN32_WINNT_WIN6 = 0x0600; // Windows Vista
+    short WIN32_WINNT_VISTA = 0x0600; // Windows Vista
+    short WIN32_WINNT_WS08 = 0x0600; // Windows Server 2008
+    short WIN32_WINNT_LONGHORN = 0x0600; // Windows Vista
+    short WIN32_WINNT_WIN7 = 0x0601; // Windows 7
+    short WIN32_WINNT_WIN8 = 0x0602; // Windows 8
+    short WIN32_WINNT_WINBLUE = 0x0603; // Windows 8.1
+    short WIN32_WINNT_WINTHRESHOLD = 0x0A00; // Windows 10
+    short WIN32_WINNT_WIN10 = 0x0A00; // Windows 10
+
     /**
      * Read the records sequentially. If this is the first read operation, the
      * EVENTLOG_FORWARDS_READ EVENTLOG_BACKWARDS_READ flags determines which